refait

+++ back/fighter/src/main/java/com/fighter/brain/Strategy.java
@@ -33,12 +33,12 @@ public class Strategy {
 		
 		boolean res = false;
 		
-		int max = -1;
+		float max = -1;
 		Vehicule vehiculeMax = null;
 		
 		for ( Vehicule vehicule : LogMission.ownVehicule ) {
 			Path path = new Path( vehicule.getVehiculeDto().getLon(), vehicule.getVehiculeDto().getLat(), this.feu.getLon(), this.feu.getLat());
-			int score = vehicule.calculScore(this.feu, vehicule.findFacility(), path);
+			float score = vehicule.calculScore(this.feu, vehicule.findFacility(), path);
 			//System.out.println("Score :" + score);
 			//System.out.println("Busy :" + vehicule.getMission() + " " + vehicule);
